add validation

+++ index.js
@@ -2,16 +2,34 @@ require('dotenv').config();
 const express = require('express');
 const superagent = require('superagent');
 const bodyParser = require('body-parser');
+const validator = require('validator');
 const db = require('./db');
 const ClientError = require('./error');
 const app = express();
 const PORT = process.env.PORT || 3000;
 const jsonParser = bodyParser.json();
 
+const MSG_PARAM = 'Missing parameters';
+const MSG_DATE = 'Invalid dates';
+const MSG_UUID = 'Invalid UUID';
+
 function asyncCall(callback) {
     return function (req, res, next) {
         callback(req, res, next).catch(next);
     }
+}
+
+function checkValidDates(startDate, endDate) {
+    const dateOptions = {
+        format: 'YYYY-MM-DD',
+        strictMode: true,
+        delimeters: ['-']
+    };
+    const isOkStart = validator.isDate(startDate, dateOptions);
+    const isOkEnd = validator.isDate(endDate, dateOptions);
+    const isOkRange = validator.isAfter(endDate, startDate);
+
+    return isOkStart && isOkEnd && isOkRange;
 }
 
 function startServer() {
@@ -22,14 +40,48 @@ function startServer() {
     app.get('/rooms/free', jsonParser, asyncCall(async (req, res) => {
         const { startDate, endDate, clientId } = req.body;
 
+        if (!startDate || !endDate || !clientId) {
+            throw new ClientError(MSG_PARAM);
+        }
+
+        if (!checkValidDates(startDate, endDate)) {
+            throw new ClientError(MSG_DATE);
+        }
+
+        if (!validator.isUUID(clientId, 4)) {
+            throw new ClientError(MSG_UUID);
+        }
+
         if (!await db.checkClientExists(clientId)) {
             throw new ClientError(`Client ${clientId} does not exist`);
         }
+
         res.send(await db.findFreeRooms(startDate, endDate, clientId));
     }));
 
     app.post('/reserve', jsonParser, asyncCall(async (req, res) => {
         const { startDate, endDate, roomId, clientId } = req.body;
+        
+        if (!startDate || !endDate || !roomId || !clientId) {
+            throw new ClientError(MSG_PARAM);
+        }
+
+        if (!checkValidDates(startDate, endDate)) {
+            throw new ClientError(MSG_DATE);
+        }
+
+        if (!validator.isUUID(roomId, 4) || !validator.isUUID(clientId, 4)) {
+            throw new ClientError(MSG_UUID);
+        }
+
+        if (!await db.checkRoomExists(roomId)) {
+            throw new ClientError(`Room ${roomId} does not exist`);
+        }
+
+        if (!await db.checkClientExists(clientId)) {
+            throw new ClientError(`Client ${clientId} does not exist`);
+        }
+        
         let isVip = false;
 
         try {
@@ -46,6 +98,15 @@ function startServer() {
 
     app.delete('/reserve/:id', asyncCall(async (req, res) => {
         const { id } = req.params;
+        
+        if (!id) {
+            throw new ClientError(MSG_PARAM);
+        }
+
+        if (!validator.isUUID(id, 4)) {
+            throw new ClientError(MSG_UUID);
+        }
+
         const isDeleted = await db.deleteReserve(id);
         let status = 200;
         let message = `Reservation ${id} was removed`;
@@ -55,6 +116,10 @@ function startServer() {
         }
         res.status(status).send(message);
     }));
+
+    app.use((req, res) => {
+        res.status(404).send('Not Found');
+    });
 
     app.use((err, req, res, next) => {
         let status = 500;
